requirements.txt
numpy
EMD_signal
ewtpy
matplotlib
padasip==1.2.2
pandas
pytz
PyWavelets
scikit-learn
#--no-binary scikit-learn
sktime
scipy
#--no-binary scipy
seaborn
#statsmodels==0.13.5
tslearn==0.6.1
vmdpy==0.2
tftb
tqdm
ssqueezepy
numba
jupyter
h5py
chart_studio
plotly
influxdb-client
paho-mqtt
#customtkinter
streamlit
streamlit-option_menu
streamlit-pdf-viewer
